Refactor ID key types to separate module
