+++ satpy/dataset/__init__.py
@@ -19,5 +19,6 @@
 
 from .anc_vars import dataset_walker, replace_anc  # noqa
 from .data_dict import DatasetDict, get_key  # noqa
-from .dataid import DataID, DataQuery, ModifierTuple, WavelengthRange, create_filtered_query  # noqa
+from .dataid import DataID, DataQuery, create_filtered_query  # noqa
+from .id_keys import ModifierTuple, WavelengthRange  # noqa
 from .metadata import combine_metadata  # noqa
