+++ satpy/etc/readers/msi_safe_l2a.yaml
@@ -12,7 +12,7 @@ reader:
     name:
       required: true
     wavelength:
-      type: !!python/name:satpy.dataset.dataid.WavelengthRange
+      type: !!python/name:satpy.dataset.id_keys.WavelengthRange
     resolution:
       transitive: false
     calibration:
