+++ satpy/etc/readers/slstr_l1b.yaml
@@ -13,7 +13,7 @@ reader:
     name:
       required: true
     wavelength:
-      type: !!python/name:satpy.dataset.dataid.WavelengthRange
+      type: !!python/name:satpy.dataset.id_keys.WavelengthRange
     resolution:
       transitive: false
     calibration:
