+++ satpy/tests/reader_tests/test_slstr_l1b.py
@@ -26,7 +26,8 @@ import numpy as np
 import pytest
 import xarray as xr
 
-from satpy.dataset.dataid import DataID, ModifierTuple, WavelengthRange
+from satpy.dataset import ModifierTuple, WavelengthRange
+from satpy.dataset.dataid import DataID
 from satpy.readers.slstr_l1b import NCSLSTR1B, NCSLSTRAngles, NCSLSTRFlag, NCSLSTRGeo
 
 local_id_keys_config = {"name": {
