+++ satpy/tests/scene_tests/test_data_access.py
@@ -22,7 +22,7 @@ import xarray as xr
 from dask import array as da
 
 from satpy import Scene
-from satpy.dataset.dataid import default_id_keys_config
+from satpy.dataset.id_keys import default_id_keys_config
 from satpy.tests.utils import FAKE_FILEHANDLER_END, FAKE_FILEHANDLER_START, make_cid, make_dataid
 
 # NOTE:
